Extract HealType from PCPCheckAndHeal
