+++ compiler/src/dotty/tools/dotc/staging/PCPCheckAndHeal.scala
@@ -2,7 +2,7 @@ package dotty.tools.dotc
 package staging
 
 import dotty.tools.dotc.ast.{tpd, untpd}
-import dotty.tools.dotc.core.Annotations.BodyAnnotation
+import dotty.tools.dotc.core.Annotations._
 import dotty.tools.dotc.core.Contexts._
 import dotty.tools.dotc.core.Decorators._
 import dotty.tools.dotc.core.Flags._
@@ -11,15 +11,10 @@ import dotty.tools.dotc.core.StagingContext._
 import dotty.tools.dotc.core.StdNames._
 import dotty.tools.dotc.core.Symbols._
 import dotty.tools.dotc.core.Types._
-import dotty.tools.dotc.util.SrcPos
-import dotty.tools.dotc.util.Spans._
-import dotty.tools.dotc.transform.SymUtils._
-import dotty.tools.dotc.typer.Checking
-import dotty.tools.dotc.typer.Implicits.SearchFailureType
-import dotty.tools.dotc.core.Annotations._
 import dotty.tools.dotc.staging.StagingLevel.*
-
 import dotty.tools.dotc.util.Property
+import dotty.tools.dotc.util.Spans._
+import dotty.tools.dotc.util.SrcPos
 
 /** Checks that the Phase Consistency Principle (PCP) holds and heals types.
  *
@@ -203,42 +198,8 @@ class PCPCheckAndHeal extends TreeMapWithStages {
    *  The tag is generated by an instance of `QuoteTypeTags` directly if the splice is explicit
    *  or indirectly by `tryHeal`.
    */
-  private def healType(pos: SrcPos)(using Context) = new TypeMap {
-    def apply(tp: Type): Type =
-      tp match
-        case tp: TypeRef =>
-          healTypeRef(tp)
-        case tp @ TermRef(NoPrefix, _) if !tp.symbol.isStatic && level > levelOf(tp.symbol) =>
-          levelError(tp.symbol, tp, pos)
-        case tp: AnnotatedType =>
-          derivedAnnotatedType(tp, apply(tp.parent), tp.annot)
-        case _ =>
-          mapOver(tp)
-
-    private def healTypeRef(tp: TypeRef): Type =
-      tp.prefix match
-        case prefix: TermRef if tp.symbol.isTypeSplice =>
-          checkNotWildcardSplice(tp)
-          if level == 0 then tp else getQuoteTypeTags.getTagRef(prefix)
-        case prefix: TermRef if !prefix.symbol.isStatic && level > levelOf(prefix.symbol) =>
-          dealiasAndTryHeal(prefix.symbol, tp, pos)
-        case NoPrefix if level > levelOf(tp.symbol) && !tp.typeSymbol.hasAnnotation(defn.QuotedRuntime_SplicedTypeAnnot) =>
-          dealiasAndTryHeal(tp.symbol, tp, pos)
-        case prefix: ThisType if level > levelOf(prefix.cls) && !tp.symbol.isStatic =>
-          dealiasAndTryHeal(tp.symbol, tp, pos)
-        case _ =>
-          mapOver(tp)
-
-    private def dealiasAndTryHeal(sym: Symbol, tp: TypeRef, pos: SrcPos)(using Context): Type =
-      val tp1 = tp.dealias
-      if tp1 != tp then apply(tp1)
-      else tryHeal(tp.symbol, tp, pos)
-
-    private def checkNotWildcardSplice(splice: TypeRef)(using Context): Unit =
-      splice.prefix.termSymbol.info.argInfos match
-        case (tb: TypeBounds) :: _ => report.error(em"Cannot splice $splice because it is a wildcard type", pos)
-        case _ =>
-  }
+  protected def healType(pos: SrcPos)(using Context) =
+    new HealType(pos)
 
   /** Check level consistency of terms references */
   private def checkLevelConsistency(tree: Ident | This)(using Context): Unit =
@@ -257,33 +218,6 @@ class PCPCheckAndHeal extends TreeMapWithStages {
              traverseChildren(tp)
     }.traverse(tree.tpe)
 
-  /** Try to heal reference to type `T` used in a higher level than its definition.
-   *  Returns a reference to a type tag generated by `QuoteTypeTags` that contains a
-   *  reference to a type alias containing the equivalent of `${summon[quoted.Type[T]]}`.
-   *  Emits and error if `T` cannot be healed and returns `T`.
-   */
-  protected def tryHeal(sym: Symbol, tp: TypeRef, pos: SrcPos)(using Context): TypeRef = {
-    val reqType = defn.QuotedTypeClass.typeRef.appliedTo(tp)
-    val tag = ctx.typer.inferImplicitArg(reqType, pos.span)
-    tag.tpe match
-
-      case tp: TermRef =>
-        ctx.typer.checkStable(tp, pos, "type witness")
-        getQuoteTypeTags.getTagRef(tp)
-      case _: SearchFailureType =>
-        report.error(
-          ctx.typer.missingArgMsg(tag, reqType, "")
-            .prepend(i"Reference to $tp within quotes requires a given $reqType in scope.\n")
-            .append("\n"),
-            pos)
-        tp
-      case _ =>
-        report.error(em"""Reference to $tp within quotes requires a given $reqType in scope.
-                     |
-                     |""", pos)
-        tp
-  }
-
   private def levelError(sym: Symbol, tp: Type, pos: SrcPos)(using Context): tp.type = {
     def symStr =
       if (!tp.isInstanceOf[ThisType]) sym.show
@@ -301,7 +235,6 @@ class PCPCheckAndHeal extends TreeMapWithStages {
           | - but the access is at level $level.$hint""", pos)
     tp
   }
-
 }
 
 object PCPCheckAndHeal {
