+++ compiler/src/dotty/tools/dotc/staging/StagingLevel.scala
@@ -1,12 +1,14 @@
-package dotty.tools.dotc.staging
+package dotty.tools.dotc
+package staging
 
-import dotty.tools.dotc.ast.{TreeMapWithImplicits, tpd}
-import dotty.tools.dotc.config.Printers.staging
-import dotty.tools.dotc.core.Decorators._
 import dotty.tools.dotc.core.Contexts._
+import dotty.tools.dotc.core.Decorators._
+import dotty.tools.dotc.core.Flags._
 import dotty.tools.dotc.core.StagingContext._
 import dotty.tools.dotc.core.Symbols._
+import dotty.tools.dotc.core.Types._
 import dotty.tools.dotc.util.Property
+import dotty.tools.dotc.util.SrcPos
 
 import scala.collection.mutable
